# Optimize the solver implementation and reduce allocations from ~1.15-1.20MB to ~38KB