--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,0 +1,10 @@
+# SarmFit 0.0.3
+Improved performance by minimizing allocations (#1) @neon-sunset
+
+# SwarmFit 0.0.2
+_Published to NuGet on 2024-09-14_
+* Published to test automated deployment
+
+# SwarmFit 0.0.1
+_Published to NuGet on 2024-09-14_
+* Initial Release