--- a/src/SwarmFit.Benchmarks/Program.cs
+++ b/src/SwarmFit.Benchmarks/Program.cs
@@ -1,0 +1,5 @@
+﻿using BenchmarkDotNet.Running;
+
+BenchmarkSwitcher
+    .FromAssembly(typeof(Program).Assembly)
+    .Run(args);