--- a/src/SwarmFit.Demo/SwarmFit.Demo.csproj
+++ b/src/SwarmFit.Demo/SwarmFit.Demo.csproj
@@ -7,6 +7,7 @@
         <UseWindowsForms>true</UseWindowsForms>
         <ImplicitUsings>enable</ImplicitUsings>
         <EnableWindowsTargeting>true</EnableWindowsTargeting>
+        <NoWarn>NU1701</NoWarn>
     </PropertyGroup>
 
     <ItemGroup>