--- a/src/SwarmFit.sln
+++ b/src/SwarmFit.sln
@@ -9,6 +9,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SwarmFit.Demo", "SwarmFit.D
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SwarmFit.Tests", "SwarmFit.Tests\SwarmFit.Tests.csproj", "{5B93F4D3-B9F2-495A-830E-63CB96A3A699}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SwarmFit.Benchmarks", "SwarmFit.Benchmarks\SwarmFit.Benchmarks.csproj", "{0F1DA296-6DFC-4C07-9172-2EAB50CA6A45}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{5B93F4D3-B9F2-495A-830E-63CB96A3A699}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5B93F4D3-B9F2-495A-830E-63CB96A3A699}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5B93F4D3-B9F2-495A-830E-63CB96A3A699}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0F1DA296-6DFC-4C07-9172-2EAB50CA6A45}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0F1DA296-6DFC-4C07-9172-2EAB50CA6A45}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0F1DA296-6DFC-4C07-9172-2EAB50CA6A45}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0F1DA296-6DFC-4C07-9172-2EAB50CA6A45}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE