--- a/src/SwarmFit/IRandomNumberGenerator.cs
+++ b/src/SwarmFit/IRandomNumberGenerator.cs
@@ -1,0 +1,9 @@
+﻿namespace SwarmFit;
+
+public interface IRandomNumberGenerator
+{
+    /// <summary>
+    /// Return a random double between 0 and 1
+    /// </summary>
+    public double NextDouble();
+}