--- a/src/SwarmFit/RandomNumberGenerators/CryptoRandom.cs
+++ b/src/SwarmFit/RandomNumberGenerators/CryptoRandom.cs
@@ -1,0 +1,14 @@
+﻿namespace SwarmFit.RandomNumberGenerators;
+
+internal class CruptoRandom : IRandomNumberGenerator
+{
+    private readonly System.Security.Cryptography.RandomNumberGenerator Rand = System.Security.Cryptography.RandomNumberGenerator.Create();
+
+    readonly byte[] bytes = new byte[sizeof(int)];
+
+    public double NextDouble()
+    {
+        Rand.GetBytes(bytes);
+        return BitConverter.ToInt32(bytes, 0) & (int.MaxValue - 1);
+    }
+}