--- a/src/SwarmFit/RandomNumberGenerators/XorShift.cs
+++ b/src/SwarmFit/RandomNumberGenerators/XorShift.cs
@@ -1,10 +1,10 @@
 ﻿namespace SwarmFit.RandomNumberGenerators;
 
+// See: https://en.wikipedia.org/wiki/Xorshift
+// Note: 2463534242 in binary is 10010010110101101000110010100010 which has a nice balance of 1s and 0s
 public class XorShift(uint seed = 2463534242) : IRandomNumberGenerator
 {
-    // note: 2463534242 in binary is 10010010110101101000110010100010 which has a nice balance of 1s and 0s
-
-    private uint Rand = seed != 0 ? seed : throw new InvalidOperationException("Don't seed me with a 0 or you'll always get 0 back");
+    private uint Rand = seed;
 
     public uint Next()
     {