--- a/src/SwarmFit/SwarmFit.csproj
+++ b/src/SwarmFit/SwarmFit.csproj
@@ -38,7 +38,6 @@
     </ItemGroup>
 
     <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
-        <PackageReference Include="System.Buffers" Version="4.5.1" />
         <PackageReference Include="System.Memory" Version="4.5.5" />
     </ItemGroup>