… README.md
… src/SwarmFit/nuget-readme.md