### Description

A friend of mine linked me this project, and I took a quick stab at optimizing the implementation to avoid unnecessary allocations that would significantly impact end user applications. Such GC impact is usually not well seen on microbenchmarks like the ones ran by BenchmarkDotNet since it invokes GC in-between the runs. Once I get to my Windows PC, I plan to see how this impacts the WinForms demo but it should be an improvement regardless.

I have changed the project to be multi-targeted in order for the collection expressions (like `[..items]`) to be able to use more efficient underlying APIs - TFM affects the Roslyn output, and NS2.0 is a really conservative target that prevents it from using any new API which is why it is recommended to provide both new and compat TFM package versions, if you want to do the latter in the first place. Adding `System.Memory` polyfill package for `Span<T>` allowed using `span.CopyTo(dest)` instead of Array's own `.CopyTo(dest, idx)` - the reason for this is array's one incurs the cost of C#'s original sin - array covariance and checks associated with it. Because spans do not allow this, their version of CopyTo is a simple length comparison and immediate call to the underlying CoreLib's `Memmove` implementation, something that array methods take extra steps to call into.

There are other small changes like ensuring that array references are not re-read per iteration from solver class fields, and are properly hoisted to locals to enable .NET to elide bound checks against them and clone the loops to provide bounds-check-less fast paths where the length bound cannot be proven. It's part of the reason why `Particle` is changed to be a struct and addressed via `ref` instead, another part is it improves data locality.

Last but not least, the biggest bottleneck here is the default *and* legacy `Random` implementation. When `Random` instance is constructed without a seed, a newer Xoshiro implementation is used which is significantly faster and has stable execution time (whatever happens with legacy one makes the results regress as they continue to get executed, and tends to cause massive bimodal gap between run latencies on Windows, but that's not just with Random, only that it exacerbates it). Unfortunately, it results in non-deterministic results with larger output variance, which makes tests crash unless the result precision tolerance is significantly relaxed, so I have only left the comment to consider this change and included reference numbers below.

In order to speed-up the algorithm further, it may be worth exploring custom RNG implementations that produce either more favourable distribution of values to this algorithm, or are faster, not being subject to the exact output the default Random with seed is limited to due to backwards compatibility considerations.

### Numbers

Environment
```
// * Summary *

BenchmarkDotNet v0.14.0, macOS Sequoia 15.0 (24A335) [Darwin 24.0.0]
Apple M1 Pro, 1 CPU, 8 logical and 8 physical cores
.NET SDK 9.0.100-rc.1.24414.7
  [Host]     : .NET 8.0.8 (8.0.824.36612), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 8.0.8 (8.0.824.36612), Arm64 RyuJIT AdvSIMD
```

Before the change

| Method    | Mean     | Error   | StdDev  | Gen0     | Allocated |
|---------- |---------:|--------:|--------:|---------:|----------:|
| SolveDemo | 104.7 ms | 0.20 ms | 0.18 ms | 166.6667 |   1.15 MB |

Note: starts at ~90ms and then regresses

After the change (with seedless but non-deterministic Random impl. provided by `new()`, makes tests unhappy with much more significant output variation)

| Method    | Mean     | Error    | StdDev   | Allocated |
|---------- |---------:|---------:|---------:|----------:|
| SolveDemo | 65.51 ms | 0.139 ms | 0.123 ms |  37.93 KB |

After the change (with current Random impl. provided by new(0))


| Method    | Mean     | Error    | StdDev   | Allocated |
|---------- |---------:|---------:|---------:|----------:|
| SolveDemo | 98.67 ms | 0.158 ms | 0.132 ms |  38.17 KB |

Note: starts at ~77ms and then regresses