@swharden 2024-09-23
Hi @neon-sunset, thanks for these optimizations! I originally wrote this code just to be educational so there are lots of allocations and unnecessary copying for the sake of simplicity. At first glace it looks like this PR improves many of them!

I may pull some features out of `SwarmFitter` and move them into a `LoggingSwarmFitter` that is built for interactivity (inspecting fits along the way, allowing users to manually request additional iterations, randomization etc.)

I'll work on this PR soon (probably tonight), merge, and publish an updated package in good time. 

Thanks again for your input! This is looking fantastic 🚀