@neon-sunset 2024-09-23
Thank you too. Honestly, on low iteration counts with the demo WinForms applications there isn't even any significant impact since all of it runs under 1ms once the solver body hits Tier1 PGO-optimized compilation. I mostly care that if there are end users that push this library much further, their experience w.r.t memory usage and GC pause frequency could be significantly improved. Feel free to leave out the changes you consider not fit for this implementation, e.g. I avoided removing LINQ in parts that were not hot or not sufficiently impactful. The main two sources of allocations are documented so if only those are applied to this, or a similar implementation elsewhere - that's fine by me as well.

Cheers!