@swharden 2024-09-24
> I mostly care that if there are end users that push this library much further

I'm on board! I wondered how many iterations it would take to solve equations with lots of unknown parameters. If I've learned anything over the years, it's that you can never predict the incredible things users will do with packages they find 🙃 

Note that I'm about to start committing a lot of things to this PR so it may seem pretty large, but I'll try to keep it simple with descriptive commit messages then do a non-squashing merge at the end.

I'm excited to see how far I can get this tonight! 🚀