@swharden 2024-09-24
@neon-sunset I'm going to merge this now but keep working in the main branch.

You may wish to take a peek at my implementation of the `IRandomNumberGenerator` and the (what I think is performant?) default.

I'm surprised that it doesn't seem to be bit-perfect deterministic, in that the "best fit" from the test system seems to be slightly different on my local computer vs. the linux computer running tests in the CI system. I'm not sure if it's a floating point error accumulation thing or something else. I'm inclined to let it go, but if you want to go down that rabbit hole I'd love to know why I had to do this 189e831b430f4e87a0731a2a28b5b54e5d577ec6 to get the tests to pass in the CI system 🤷