@swharden 2024-09-24
Hi @neon-sunset, I made some good progress but am going to put this down tonight.

| Method    | Mean     | Error     | StdDev    | Allocated |
|---------- |---------:|----------:|----------:|----------:|
| SolveDemo | 2.347 ms | 0.0837 ms | 0.0046 ms |  26.72 KB |

Check out the demo app to see how I changed the behavior to support successive iterative improvement.

I think the benchmark should be more stable now, so let me know if you have ideas for further improvements and we can open up a new PR if you're still interested in diving deeper! Thanks again for your help on this. I'll probably work on this another day or so before publishing an updated package.

![fit12](https://github.com/user-attachments/assets/6fddd84c-45d6-44e8-9234-ac80bb1cb449)